Get the key from the user system environment variable

Saved in CAX_KEY in the .Renviron file. Alternatively it can be passed in.

check_key(key)

Arguments

key

the key

References

This function is modeled off the functions in https://github.com/ropensci/rredlist