Internal data used by functions

rCAX has a number of internally stored data (in sysdata.rda) that are used by the functions. If the user wants to use these, preface the name with rCAX:::. For example, rCAX:::caxesu shows the list of ESU/DPS names.

Details

Internal data

caxsuperpops: the SuperPopulations table

caxtabs: the API table names and their table_ids. Used to look up table_id

caxpops: the Populations table

caxesu: the list of ESU/DPS names. Useful for filtering.

nosa_xport_colnames, sar_xport_colnames, pni_xport_colnames, rpers_xport_colnames, presmolt_xport_colnames, juvout_xport_colnames: the column names that appear in the HLI Tabular Query output so that rCAX _xport functions produce the same columns in the same order.

cax_column_definitions: a data frame with column name and definitions. Produced from the HLI Tabular Query Excel files

Examples

rCAX:::caxesu
#>  [1] "N/A"                                                      
#>  [2] "Salmon, Chinook (Lower Columbia River ESU)"               
#>  [3] "Salmon, Chinook (Puget Sound ESU)"                        
#>  [4] "Salmon, Chinook (Snake River fall-run ESU)"               
#>  [5] "Salmon, Chinook (Snake River spring/summer-run ESU)"      
#>  [6] "Salmon, Chinook (Upper Columbia River spring-run ESU--XN)"
#>  [7] "Salmon, Chinook (Upper Columbia River spring-run ESU)"    
#>  [8] "Salmon, Chinook (Upper Willamette River ESU)"             
#>  [9] "Salmon, chum (Columbia River ESU)"                        
#> [10] "Salmon, chum (Hood Canal summer-run ESU)"                 
#> [11] "Salmon, coho (Lower Columbia River ESU)"                  
#> [12] "Salmon, coho (Oregon Coast ESU)"                          
#> [13] "Salmon, sockeye (Ozette Lake ESU)"                        
#> [14] "Salmon, sockeye (Snake River ESU)"                        
#> [15] "Steelhead (Lower Columbia River DPS)"                     
#> [16] "Steelhead (Middle Columbia River DPS--XN)"                
#> [17] "Steelhead (Middle Columbia River DPS)"                    
#> [18] "Steelhead (Puget Sound DPS)"                              
#> [19] "Steelhead (Snake River Basin DPS)"                        
#> [20] "Steelhead (Upper Columbia River DPS)"                     
#> [21] "Steelhead (Upper Willamette River DPS)"