Get EscData table

Returns EscData table sorted by refid and countdate.

The EscData table_id is "F11B5228-F716-487B-807D-0DD04B96EEB8" and is set automatically using a saved data frame from a rcax_tables() call. The table is saved in R/sysdata.rda.

rcax_escdata() will download 1000 records.

rcax_escdata(
  tablename = "EscData",
  flist = NULL,
  qlist = NULL,
  cols = NULL,
  sortcols = c("countdate", "refid"),
  type = c("data.frame", "colnames"),
  GETargs = list(table_id = NULL, recordloc = "records", key = NULL, parse = TRUE),
  ...
)

Arguments

tablename

(character) The CAX name for the table; ignored if GETargs$table_id passed in.

flist

A filter for the query. See details and rcax_filter()

qlist

Additional query params. See rcax_table_query() for details.

cols

column names to return. Use cols=NULL if you want all columns

sortcols

The columns to sort on for the returned table. The order of sortcols indicates the order of sorting. Start with the smallest group and work outwards, e.g. c("spawningyear", "popid", "MPG")

type

whether to return the table ("data.frame") or colnames ("colnames"). Default is to return the table as a data frame.

GETargs

A list of arguments for the rcax_GET() call. These do not normally need to be specified by the user. table_id This is the CAX table id. It is looked up using tablename and caxpops. recordloc This is the name of the record we want in the list returned from the GET call. key is the API key. See rcax_GET() for API key details.

Value

data frame if type="data.frame" (default) and the colnames if type="colnames"

References

API docs at https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/

Examples

rcax_escdata(cols = NULL, qlist=list(limit=5))
#>   ascode countpermile repeatspawners asnid agencyid countcidisttype
#> 2     98            0                   98       10                
#> 3     98                                98       15                
#> 5     98                                98       15                
#> 1     98                                98       15                
#> 4     98                                98       36                
#>   countciuplim compilerrecordid                   begindate
#> 2                               September, 03 2002 00:00:00
#> 3                                   March, 01 1973 00:00:00
#> 5                                 October, 01 1961 00:00:00
#> 1                                November, 01 1979 00:00:00
#> 4                                   March, 15 1962 00:00:00
#>                                     id                   upddate num
#> 2 0000219F-A662-4A16-8F46-ED171C56990A   March, 27 2006 00:00:00   2
#> 3 0000578E-E906-41CA-A778-7EC7A8C06DDA     May, 23 2014 00:00:00   3
#> 5 0001112D-987D-4F27-88B3-EF027074AE6A January, 18 2000 16:53:00   5
#> 1 00000A70-075A-4AE1-857D-42BC1048939C    June, 01 2011 00:00:00   1
#> 4 00010D89-9D97-44FD-A9F5-D55C75C4EC6E  August, 16 2010 00:00:00   4
#>                dataentry
#> 2         IDFG StreamNet
#> 3   April Waters, ed JLC
#> 5        Briana Sounhein
#> 1 Disk 97-12, ed 0611 KG
#> 4               cshannon
#>                                                                                                                                                                                                                          countcom
#> 2                                                                                                                                                                                                                                
#> 3                                                                                                                                                                                                       No counts made this year.
#> 5                                                                                                                                                                                                                         Table 1
#> 1                                                                                                                                                                                                                                
#> 4 No mainstem estimate was developed and a grand total for spring-run spawning was not made.  Tributary estimates were Mill Creek, 2000, Big Chico Creek, 200, Butte Creek, 2000 and there was no estimate for the Feather River.
#>   publish                   countdate refid sampmethid
#> 2     Yes September, 03 2002 00:00:00 46758        213
#> 3     Yes                             50038         98
#> 5     Yes                             50061        105
#> 1     Yes                             53391        203
#> 4     Yes                             90001         98
#>                       enddate calcmethid countvalue trendid timessurveyed
#> 2 September, 03 2002 00:00:00        304          0   44227             1
#> 3   October, 31 1973 00:00:00         98              53571              
#> 5   January, 31 1962 00:00:00        304          3   50321              
#> 1   October, 31 1980 00:00:00        314       1024   54227              
#> 4   October, 14 1962 00:00:00         98              90661              
#>   asmethod assource countcilowlim milessurveyed countcilevel nullflag
#> 2       98                            2.9820075                    No
#> 3       98                                                        Yes
#> 5       98                                  0.3                    No
#> 1       98                                                         No
#> 4       98                                                        Yes

# to print the first 5 column names
rcax_escdata(type="colnames")[1:5]
#> [1] "ascode"         "countpermile"   "repeatspawners" "asnid"         
#> [5] "agencyid"