Query the NOSA table

Returns NOSA table sorted by popid and spawningyear. The default columns to return are "recoverydomain", "esu_dps", "commonpopname", "commonname", "run", "popid", "majorpopgroup", "spawningyear", "tsaij", "tsaej", "nosaij", "nosaej"

rcax_nosa() will download 1000 records from the NOSA table. You will probably want to make a filtered query by passing in value to filter on via flist. See rcax_filter(). Examples for the NOSA table:

rcax_nosa(flist=list(popid=7) return values for popid 7.

rcax_nosa(flist=list(esu_dps="Salmon, Chinook (Snake River spring/summer-run ESU)") return values for one ESU. Use rCAX:::caxesu to see a list of ESU_DPS names.

rcax_nosa(
  tablename = "NOSA",
  flist = NULL,
  qlist = NULL,
  cols = c("recoverydomain", "esu_dps", "commonpopname", "commonname", "run", "popid",
    "majorpopgroup", "spawningyear", "tsaij", "tsaej", "nosaij", "nosaej"),
  sortcols = c("spawningyear", "popid"),
  type = c("data.frame", "colnames"),
  GETargs = list(table_id = NULL, recordloc = "records", key = NULL, parse = TRUE),
  ...
)

Arguments

tablename

(character) The CAX name for the table; ignored if GETargs$table_id passed in.

flist

A filter for the query. See details and rcax_filter()

qlist

Additional query params. See rcax_table_query() for details.

cols

column names to return. Use cols=NULL if you want all columns

sortcols

The columns to sort on for the returned table. The order of sortcols indicates the order of sorting. Start with the smallest group and work outwards, e.g. c("spawningyear", "popid", "MPG")

type

whether to return the table ("data.frame") or colnames ("colnames"). Default is to return the table as a data frame.

GETargs

A list of arguments for the rcax_GET() call. These do not normally need to be specified by the user. table_id This is the CAX table id. It is looked up using tablename and caxpops. recordloc This is the name of the record we want in the list returned from the GET call. key is the API key. See rcax_GET() for API key details.

Value

data frame if type="data.frame" (default) and the colnames if type="colnames"

Details

The NOSA table_id is "4EF09E86-2AA8-4C98-A983-A272C2C2C7E3" and is set automatically using a saved data frame from a rcax_tables() call. The table is saved in R/sysdata.rda.

References

API docs at https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/

See also

Examples

a <- rcax_nosa(flist=list(popid=7))
a[, c("popid", "spawningyear", "nosaij", "nosaej")] |> head()
#>    popid spawningyear nosaij nosaej
#> 31     7         1953   4273   3484
#> 46     7         1954    973    816
#> 16     7         1955    122    106
#> 17     7         1956   2606   2338
#> 20     7         1957   1754   1619
#> 13     7         1958    486    456

rcax_nosa(qlist=list(limit=1), cols=NULL)
#>   nosaijupperlimit tsaej age8propupperlimit metriclocation popfitnotes
#> 1                                                                     
#>                                            cbfwapopname age9proplowerlimit
#> 1 Deschutes River Eastside Tributaries Summer Steelhead                   
#>   phosejalpha contactpersonfirst compilerrecordid
#> 1                          Kasey      30618010304
#>                                     id age3propupperlimit nosaejlowerlimit
#> 1 0002B923-F495-401B-BFB6-685F17D78FF2                                    
#>   nosjf phosej nosjflowerlimit popfit spawningyear contactpersonlast
#> 1                                Same         2001          Bliesner
#>       dataentry age10propupperlimit age4proplowerlimit publish nosaej age3prop
#> 1 Jake Chambers                                            Yes            0.07
#>   escapementlat age6propupperlimit phosejupperlimit
#> 1                                                  
#>                                          waterbody
#> 1 Bakeoven Creek, Buckhollow Creek and Trout Creek
#>                                                                                                                                               comments
#> 1 Proportion at age determined from average total age of scales collected from wild Deschutes fish in the mainstem, 1970 and 1971 (Olsen et al. 1991).
#>   age5prop age3proplowerlimit age2proplowerlimit
#> 1      0.3                                      
#>                        majorpopgroup otherdatasources tsaejalpha
#> 1 Cascades Eastern Slope Tributaries                            
#>   age5propupperlimit nosaijalpha age10proplowerlimit nosaijlowerlimit
#> 1                                                                    
#>   nosjfupperlimit phosij num phosejlowerlimit age7prop    run refid hosjf
#> 1                  0.137   1                      0.01 Summer            
#>                                                          measurelocation
#> 1 Redd count data may be found at https://www.streamnet.org/data/trends/
#>   tsaijlowerlimit escapementtiming commonname contactphone age4propupperlimit
#> 1                                   Steelhead 541-975-3670                   
#>   estimatetype phosijupperlimit popid age11pluspropupperlimit tsaijupperlimit
#> 1         NOSA                     58                                        
#>   phosijalpha tsaij datastatus escapementlong tsaejlowerlimit
#> 1             10005      Final                               
#>   age5proplowerlimit nobroodstockremoved methodnumber age9prop
#> 1                                                   1         
#>                     dataentrynotes agepropalpha    recoverydomain
#> 1 ODFW - East Region Fish Research              Interior Columbia
#>   age7propupperlimit nosjfalpha nosaij phosijlowerlimit age8proplowerlimit
#> 1                                 8637                                    
#>                                                                                                            metacomments
#> 1 ODFW reserves the right to review and revise data as needed. Consult the contact person for metric, and measure data.
#>                            contactagency age2propupperlimit age2prop nullrecord
#> 1 Oregon Department of Fish and Wildlife                                     No
#>   protmethdocumentation bestvalue tsaijalpha
#> 1                             Yes           
#>                                                                           protmethurl
#> 1 https://nrimp.dfw.state.or.us/DataClearinghouse/default.aspx?p=202&XMLname=1100.xml
#>                                           indicatorlocation nosaejalpha
#> 1 http://odfwrecoverytracker.org/explorer/species/Steelhead            
#>   nosaejupperlimit submitagency age7proplowerlimit age4prop age8prop
#> 1                          ODFW                        0.57         
#>                                                                                                                                                                                                        protmethname
#> 1 A Compendium of Viable Salmonid Population Abundance and Productivity Field and Analysis Methods for Natural Origin Steelhead Populations in the Middle Columbia River DPS of Northeast Oregon from 1959 to 2015.
#>                   upddate age9propupperlimit
#> 1 April, 09 2021 15:16:51                   
#>                                 esu_dps                    methodadjustments
#> 1 Steelhead (Middle Columbia River DPS) Single Pass Index SGS Redd Expansion
#>                     contactemail age10prop age11plusprop
#> 1 kasey.bliesner@odfw.oregon.gov                        
#>   age11plusproplowerlimit            commonpopname age6prop tsaejupperlimit
#> 1                         Deschutes River eastside     0.05                
#>   age6proplowerlimit
#> 1                   

# to print the first 5 column names
rcax_nosa(type="colnames")[1:5]
#> [1] "nosaijupperlimit"   "tsaej"              "age8propupperlimit"
#> [4] "metriclocation"     "popfitnotes"