Query the NOSA XPort table

Returns NOSA with some additional metadata (such as NMFS_PopID and update dates). This is the table that one can download from the CAP Fish HLIs Tabular Query.

rcax_nosa_xport() will download 1000 records. You will want to make a filtered query by passing in values to filter on via flist. Keep in mind that the field names and values are case insensitive. Using POPID, popid and PopID will have the same effect. Examples:

rcax_nosa_xport(flist=list(popid=7) return values for popid 7.

rcax_nosa_xport(flist=list(nmfs_popid=6) return values for NMFS_PopID 6.

rcax_nosa_xport(flist=list(esu_dps="Salmon, Chinook (Snake River spring/summer-run ESU)") return values for one ESU. Use rCAX:::caxesu to see a list of ESU_DPS names.

rcax_nosa_xport(
  tablename = "XPortCA_NOSA_01",
  flist = NULL,
  qlist = NULL,
  cols = NULL,
  sortcols = c("spawningyear", "popid"),
  type = c("data.frame", "colnames"),
  GETargs = list(table_id = NULL, recordloc = "records", key = NULL, parse = TRUE),
  ...
)

Arguments

tablename

(character) The CAX name for the table; ignored if GETargs$table_id passed in.

flist

A filter for the query. See details and rcax_filter()

qlist

Additional query params. See rcax_table_query() for details.

cols

column names to return. Use cols=NULL if you want all columns

sortcols

The columns to sort on for the returned table. The order of sortcols indicates the order of sorting. Start with the smallest group and work outwards, e.g. c("spawningyear", "popid", "MPG")

type

whether to return the table ("data.frame") or colnames ("colnames"). Default is to return the table as a data frame.

GETargs

A list of arguments for the rcax_GET() call. These do not normally need to be specified by the user. table_id This is the CAX table id. It is looked up using tablename and caxpops. recordloc This is the name of the record we want in the list returned from the GET call. key is the API key. See rcax_GET() for API key details.

Value

data frame if type="data.frame" (default) and the colnames if type="colnames"

Details

The table_id is set automatically using a saved data frame from a rcax_tables() call. The table is saved in R/sysdata.rda.

The table has a few columns that do not appear in the Excel files that one can download from StreamNet: "species", "publish", "num", "hli", "agency", "esudps", "hli_id". "esudps" is a slight variant on "esu_dps" and is removed. "num" is a sort flag and is removed. "publish" will always be Yes but is left in. The other columns are left in. The colnames are resorted into the order found in the downloaded Excel files with the extra column added to the end.

References

API docs at https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/

See also

Examples

a <- rcax_nosa_xport(
         flist=list(nmfs_popid=7),
         cols=c("nmfs_popid", "spawningyear", "tsaej", "nosaej")
         )
head(a)
#>    nmfs_popid spawningyear tsaej nosaej
#> 49          7         1964  3020       
#> 50          7         1965  2539       
#> 51          7         1966  2984       
#> 48          7         1967  3182       
#> 52          7         1968  3594       
#> 5           7         1969  2973       

# First 3 columns
rcax_nosa_xport(qlist=list(limit=3))[,1:3]
#>       commonname    run    recoverydomain
#> 3 Chinook salmon Spring Interior Columbia
#> 1      Steelhead Summer Interior Columbia
#> 2    Coho salmon   Fall      Oregon Coast

# to print the first 5 column names
rcax_nosa_xport(type="colnames")[1:5]
#> [1] "nosaijupperlimit"   "tsaej"              "age8propupperlimit"
#> [4] "metriclocation"     "popfitnotes"       

# to print the first 5 column names as they appear
# in Excel files downloaded from https://www.streamnet.org/data/hli/
tab <- rcax_nosa_xport(qlist=list(limit=1))
colnames(tab)[1:5]
#> [1] "commonname"     "run"            "recoverydomain" "esu_dps"       
#> [5] "majorpopgroup"