Return PNI base table

Returns the base HLI table which is missing the additional metadata (such as NMFS_PopID and update dates) found in the XPort tables.

The query will download 1000 records by default; you can change this by passing in qlist=list(limit=1100) for example. To make a filtered query e.g. for one ESU or popid) pass in the flist argument. Field names and values are case insensitive, e.g. POPID, popid and PopID return the same result. If your filter value is not in the table, the returned table will be empty. Examples:

flist=list(popid=7) return values for popid 7.

flist=list(esu_dps="Salmon, Chinook (Snake River spring/summer-run ESU)") return values for one ESU. Use rCAX:::caxesu to see a list of ESU_DPS names.

rcax_pni(
  tablename = "PNI",
  flist = NULL,
  qlist = NULL,
  cols = NULL,
  sortcols = c("spawningyear", "popid"),
  type = c("data.frame", "colnames"),
  GETargs = list(table_id = NULL, recordloc = "records", key = NULL, parse = TRUE),
  ...
)

Arguments

tablename

(character) The CAX name for the table; ignored if GETargs$table_id passed in.

flist

A filter for the query. See details and rcax_filter()

qlist

Additional query params. See rcax_table_query() for details.

cols

column names to return. Use cols=NULL if you want all columns

sortcols

The columns to sort on for the returned table. The order of sortcols indicates the order of sorting. Start with the smallest group and work outwards, e.g. c("spawningyear", "popid", "MPG")

type

whether to return the table ("data.frame") or colnames with definitions ("colnames"). Default is to return the table as a data frame.

GETargs

A list of arguments for the rcax_GET() call. These do not normally need to be specified by the user. table_id This is the CAX table id. It is looked up using tablename and rCAX:::caxpops (an internal dataset). recordloc This is the name of the record we want in the list returned from the GET call. key is the API key. See vignette("setup", package = "rCAX") for API key details.

Value

data frame if type="data.frame" (default) and the colnames if type="colnames"

Details

The table_id is set automatically using a saved data frame from a rcax_tables() call. The table is saved in R/sysdata.rda.

If you want to see only the column names use type="colnames.

References

API docs at https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/

See also

Examples

# return popid of first record
# Note the part after $ is case sensitive
id <- rcax_pni(qlist=list(limit=1))$popid

# Return the table for this id
a <- rcax_pni(
         flist=list(popid=id),
         cols=c("popid", "spawningyear", "pniej")
         )
head(a)
#>   popid spawningyear pniej
#> 1    15         1985  1.00
#> 2    15         1986  1.00
#> 3    15         1987  1.00
#> 4    15         1988  1.00
#> 5    15         1989  0.98
#> 6    15         1990  0.69

# First 3 columns
rcax_pni(qlist=list(limit=3))[,1:3]
#>   popid                                            metriclocation
#> 1    15                                                          
#> 2    47                                                          
#> 3   512 https://docs.cbfwl.org/StreamNet_References/WAsn17274.pdf
#>                                                      popfitnotes
#> 1                                                               
#> 2 Nason Creek Spring Chinook Portion of Wenatchee Spring Chinook
#> 3                                               Whole population

# to print the first column names and definitions
head(rcax_pni(type="colnames"))
#>             name
#> 1          popid
#> 2 metriclocation
#> 3    popfitnotes
#> 4     datastatus
#> 5   cbfwapopname
#> 6   methodnumber
#>                                                                                           definition
#> 1                                                         StreamNet code for the population of fish.
#> 2                                          Where this supporting metric is maintained at the source.
#> 3                      Text description of why the data are not at the scale of a single population.
#> 4                                                                   "Draft", "Reviewed", or "Final".
#> 5                                            Population name defined by CBFWA for subbasin planning.
#> 6 When an organization provides > 1 estimate per year, this number identifies which method was used.

# to print the first 5 column names as they appear
# in Excel files downloaded from https://www.streamnet.org/data/hli/
tab <- rcax_pni(qlist=list(limit=1))
colnames(tab)[1:5]
#> [1] "popid"          "metriclocation" "popfitnotes"    "datastatus"    
#> [5] "cbfwapopname"