Return SAR table from CAP Fish HLIs Tabular Query

Returns the HLI table with some additional metadata (such as NMFS_PopID and update dates). This is the table that one can download from the CAP Fish HLIs Tabular Query.

The query will download 1000 records by default; you can change this by passing in qlist=list(limit=1100) for example. To make a filtered query e.g. for one ESU or popid) pass in the flist argument. Field names and values are case insensitive, e.g. POPID, popid and PopID return the same result. If your filter value is not in the table, the returned table will be empty. Examples:

flist=list(popid=7) return values for popid 7.

flist=list(nmfs_popid=6) return values for NMFS_PopID 6.

flist=list(esu_dps="Salmon, Chinook (Snake River spring/summer-run ESU)") return values for one ESU. Use rCAX:::caxesu to see a list of ESU_DPS names.

rcax_sar_xport(
  tablename = "XPortCA_SAR_01",
  flist = NULL,
  qlist = NULL,
  cols = NULL,
  sortcols = c("outmigrationyear", "popid"),
  type = c("data.frame", "colnames"),
  GETargs = list(table_id = NULL, recordloc = "records", key = NULL, parse = TRUE),
  ...
)

Arguments

tablename

(character) The CAX name for the table; ignored if GETargs$table_id passed in.

flist

A filter for the query. See details and rcax_filter()

qlist

Additional query params. See rcax_table_query() for details.

cols

column names to return. Use cols=NULL if you want all columns

sortcols

The columns to sort on for the returned table. The order of sortcols indicates the order of sorting. Start with the smallest group and work outwards, e.g. c("spawningyear", "popid", "MPG")

type

whether to return the table ("data.frame") or colnames ("colnames"). Default is to return the table as a data frame.

GETargs

A list of arguments for the rcax_GET() call. These do not normally need to be specified by the user. table_id This is the CAX table id. It is looked up using tablename and caxpops. recordloc This is the name of the record we want in the list returned from the GET call. key is the API key. See rcax_GET() for API key details.

Value

data frame if type="data.frame" (default) and the colnames if type="colnames"

Details

The table_id is set automatically using a saved data frame from a rcax_tables() call. The table is saved in R/sysdata.rda.

The XPORT tables have a few columns that do not appear in the Excel files that one can download from StreamNet: "species", "publish", "num", "hli", "agency", "esudps", "hli_id". "esudps" is a slight variant on "esu_dps" and is removed. "num" is a sort flag and is removed. "publish" will always be Yes but is left in. The other columns are left in. The colnames are re-sorted into the order found in the downloaded Excel files with the extra columns added to the end.

If you want to see the original columns in the original order, use type="colnames and use that output as the value for the cols argument.

References

API docs at https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/

See also

Examples

# return NMFS_PopID of first record
# Note the part after $ is case sensitive
id <- rcax_sar_xport(qlist=list(limit=1))$nmfs_popid

a <- rcax_sar_xport(
         flist=list(nmfs_popid=id),
         cols=c("nmfs_popid", "outmigrationyear", "sar")
         )
head(a)
#>   nmfs_popid outmigrationyear  sar
#> 1        251             2007 4.20
#> 2        251             2007 2.45
#> 3        251             2008 6.74
#> 4        251             2008 3.57
#> 5        251             2009 2.05
#> 6        251             2009 3.77

# First 3 columns
rcax_sar_xport(qlist=list(limit=3))[,1:3]
#>       commonname    run    recoverydomain
#> 1 Chinook salmon Spring Interior Columbia
#> 2      Steelhead Winter Interior Columbia
#> 3    Coho salmon   Fall      Oregon Coast

# to print the first 5 column names
rcax_sar_xport(type="colnames")[1:5]
#> [1] "popid"          "tsoalpha"       "metriclocation" "popfitnotes"   
#> [5] "tar"           

# to print the first 5 column names as they appear
# in Excel files downloaded from https://www.streamnet.org/data/hli/
tab <- rcax_sar_xport(qlist=list(limit=1))
colnames(tab)[1:5]
#> [1] "commonname"     "run"            "recoverydomain" "esu_dps"       
#> [5] "majorpopgroup"