Get SuperPopulations table

Returns SuperPopulations table

The SuperPopulations table_id is "009A08FE-6479-44FC-9B6F-01C55E2C8BA3" and is set automatically using a saved data frame from a rcax_tables() call. The table is saved in R/sysdata.rda.

rcax_superpops() will download 1000 records. Pass in flist to filter.

rcax_superpops(
  tablename = "SuperPopulations",
  flist = NULL,
  qlist = NULL,
  cols = NULL,
  sortcols = c("popid"),
  type = c("data.frame", "colnames"),
  GETargs = list(table_id = NULL, recordloc = "records", key = NULL, parse = TRUE),
  ...
)

Arguments

tablename

(character) The CAX name for the table; ignored if GETargs$table_id passed in.

flist

A filter for the query. See details and rcax_filter()

qlist

Additional query params. See rcax_table_query() for details.

cols

column names to return. Use cols=NULL if you want all columns

sortcols

The columns to sort on for the returned table. The order of sortcols indicates the order of sorting. Start with the smallest group and work outwards, e.g. c("spawningyear", "popid", "MPG")

type

whether to return the table ("data.frame") or colnames ("colnames"). Default is to return the table as a data frame.

GETargs

A list of arguments for the rcax_GET() call. These do not normally need to be specified by the user. table_id This is the CAX table id. It is looked up using tablename and caxpops. recordloc This is the name of the record we want in the list returned from the GET call. key is the API key. See rcax_GET() for API key details.

Value

data frame if type="data.frame" (default) and the colnames if type="colnames"

References

API docs at https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/

See also

Examples

# to get all columns and print the column names
rcax_superpops(type="colnames")
#> [1] "popid"         "popfit"        "popfitnotes"   "upddate"      
#> [5] "contactagency" "num"           "id"            "publish"      
#> [9] "superpopid"   

# get one record
rcax_superpops(qlist=list(limit=1))
#> [1] "The table includes no data. Did you specify cols that do not appear in the table? Use rcax_nosa_xport(type=\"colnames\")\" to see the colnames. Names are case insensitive."