Get EscData table

Returns EscData table sorted by refid and countdate.

The EscData table_id is "F11B5228-F716-487B-807D-0DD04B96EEB8" and is set automatically using a saved data frame from a rcax_tables() call. The table is saved in R/sysdata.rda.

rcax_escdata() will download 1000 records.

rcax_escdata(
  extra = NULL,
  table_id = NULL,
  key = NULL,
  parse = TRUE,
  cols = NULL,
  ...
)

Arguments

extra

extra query params or column values. Enter as a list, e.g. list(pageno=7)

table_id

not normally needed as this is set automatically to the table id for NOSA using the table saved from rcax_table(). But you can set it if needed.

key

A CAX API token. See rcax_key.

parse

(logical) Whether to parse to list (FALSE) or data.frame (TRUE). Default: TRUE

cols

column names to return. Use cols=NULL if you want all columns

...

Curl options passed to HttpClient

popid

(character or number) A NMFS_POPID for a population

References

API docs at https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/

This function is modeled off the functions in https://github.com/ropensci/rredlist

Examples

# to get all columns and print the column names
# a <- rcax_escdata(cols = NULL)
# colnames(a)