Get nosa table for a popid

Returns nosa table sorted by popid and spawningyear. The default columns to return are "recoverydomain", "esu_dps", "commonpopname", "commonname", "run", "popid", "majorpopgroup", "spawningyear", "nosaij", "nosaej"

The NOSA table_id is "4EF09E86-2AA8-4C98-A983-A272C2C2C7E3"

rcax_nosa() will download 1000 records from the NOSA table. You will probably want to make a filtered query by passing in value for a specific popid, e.g. rcax_nosa(popid=7)

rcax_nosa(
  table_id = "4EF09E86-2AA8-4C98-A983-A272C2C2C7E3",
  popid = NULL,
  extra = NULL,
  key = NULL,
  parse = TRUE,
  cols = c("recoverydomain", "esu_dps", "commonpopname", "commonname", "run", "popid",
    "majorpopgroup", "spawningyear", "nosaij", "nosaej"),
  ...
)

Arguments

table_id

table id eg 4EF09E86-2AA8-4C98-A983-A272C2C2C7E3

popid

a popid number

extra

extra query params or column values. Enter as a list, e.g. list(pageno=7)

key

A CAX API token. See rcax_key.

parse

(logical) Whether to parse to list (FALSE) or data.frame (TRUE). Default: TRUE

cols

column names to return. Use cols=NULL if you want all columns

...

Curl options passed to HttpClient

References

API docs at https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/

This function is modeled off the functions in https://github.com/ropensci/rredlist

Examples

a <- rcax_nosa(popid=7)
a[, c("popid", "spawningyear", "nosaij", "nosaej")]
#>    popid spawningyear nosaij nosaej
#> 31     7         1953   4273   3484
#> 46     7         1954    973    816
#> 16     7         1955    122    106
#> 17     7         1956   2606   2338
#> 20     7         1957   1754   1619
#> 13     7         1958    486    456
#> 62     7         1959    712    655
#> 26     7         1960   3161   2947
#> 34     7         1961    347    285
#> 61     7         1962   1146   1081
#> 78     7         1963    313    261
#> 71     7         1964    208    181
#> 25     7         1965    770    739
#> 50     7         1966    626    507
#> 27     7         1967   1625   1370
#> 22     7         1968    822    728
#> 45     7         1969   1882   1819
#> 18     7         1970    992    910
#> 12     7         1971   2319   2206
#> 49     7         1972   1106   1079
#> 55     7         1973   1855   1816
#> 44     7         1974    753    738
#> 70     7         1975    399    359
#> 74     7         1976   1028    973
#> 33     7         1977    104    100
#> 69     7         1978    878    878
#> 32     7         1979    579    557
#> 68     7         1980   1034    994
#> 21     7         1981    275    264
#> 63     7         1982    745    716
#> 58     7         1983    733    733
#> 24     7         1984    378    378
#> 43     7         1985    386    371
#> 60     7         1986    319    310
#> 35     7         1987    152    149
#> 23     7         1988    163    160
#> 65     7         1989     72     72
#> 76     7         1990      0      0
#> 30     7         1991      9      8
#> 77     7         1992     40     40
#> 6      7         1993    107     99
#> 41     7         1994     14     14
#> 79     7         1995     38     34
#> 51     7         1996     35     34
#> 5      7         1997     72     72
#> 28     7         1998     91     91
#> 40     7         1999     81     81
#> 19     7         2000     54     49
#> 75     7         2001    397    357
#> 64     7         2002    217    204
#> 29     7         2003    175    175
#> 36     7         2004     37     35
#> 1      7         2005     38     35
#> 15     7         2006     93     84
#> 10     7         2007     51     46
#> 14     7         2008     79     73
#> 57     7         2009    129     92
#> 8      7         2010    463    438
#> 38     7         2010    292       
#> 2      7         2011    615    540
#> 48     7         2011    152       
#> 11     7         2012    328    316
#> 73     7         2012    391       
#> 4      7         2013    511       
#> 52     7         2013    275    187
#> 3      7         2014    570    531
#> 66     7         2014    981       
#> 39     7         2015    256    242
#> 42     7         2015    309       
#> 37     7         2016    190    182
#> 47     7         2016    260       
#> 53     7         2017     72       
#> 67     7         2017     39     33
#> 9      7         2018     68       
#> 56     7         2018     31     27
#> 54     7         2019    101       
#> 59     7         2019     55     49
#> 7      7         2020    160    150
#> 72     7         2021    111    102

a <- rcax_nosa()
head(a)
#>        recoverydomain                                             esu_dps
#> 495 Interior Columbia Salmon, Chinook (Snake River spring/summer-run ESU)
#> 441 Interior Columbia Salmon, Chinook (Snake River spring/summer-run ESU)
#> 3   Interior Columbia Salmon, Chinook (Snake River spring/summer-run ESU)
#> 634 Interior Columbia Salmon, Chinook (Snake River spring/summer-run ESU)
#> 21  Interior Columbia Salmon, Chinook (Snake River spring/summer-run ESU)
#> 387 Interior Columbia Salmon, Chinook (Snake River spring/summer-run ESU)
#>                      commonpopname     commonname    run popid
#> 495 Catherine Creek spring Chinook Chinook salmon Spring     7
#> 441 Catherine Creek spring Chinook Chinook salmon Spring     7
#> 3   Catherine Creek spring Chinook Chinook salmon Spring     7
#> 634 Catherine Creek spring Chinook Chinook salmon Spring     7
#> 21  Catherine Creek spring Chinook Chinook salmon Spring     7
#> 387                          GRCAT Chinook salmon spring     7
#>           majorpopgroup spawningyear nosaij nosaej
#> 495 Grande Ronde/Imnaha         1993    107     99
#> 441 Grande Ronde/Imnaha         1997     72     72
#> 3   Grande Ronde/Imnaha         2005     38     35
#> 634 Grande Ronde/Imnaha         2010    463    438
#> 21  Grande Ronde/Imnaha         2011    615    540
#> 387 Grande Ronde/Imnaha         2013    511       

# to get all columns and print the column names
# a <- rcax_nosa(cols = NULL)
# colnames(a)