Get nosa table

Returns nosa table sorted by popid and spawningyear. The default columns to return are "recoverydomain", "esu_dps", "commonpopname", "commonname", "run", "popid", "majorpopgroup", "spawningyear", "nosaij", "nosaej"

The NOSA table is "4EF09E86-2AA8-4C98-A983-A272C2C2C7E3"

rcax_nosa(
  table_id = "4EF09E86-2AA8-4C98-A983-A272C2C2C7E3",
  extra = NULL,
  key = NULL,
  parse = TRUE,
  cols = c("recoverydomain", "esu_dps", "commonpopname", "commonname", "run", "popid",
    "majorpopgroup", "spawningyear", "nosaij", "nosaej"),
  ...
)

rcax_nosa_(table_id, extra = NULL, key = NULL, ...)

Arguments

table_id

table id eg 4EF09E86-2AA8-4C98-A983-A272C2C2C7E3

extra

extra query params. Enter as a list, e.g. list(pageno=1)

key

A CAX API token. See rcax_use_iucn.

parse

(logical) Whether to parse to list (FALSE) or data.frame (TRUE). Default: TRUE

cols

column names to return. Use cols=NULL if you want all columns

...

Curl options passed to HttpClient

Value

A list, with the data in the result slot, unless using a function with a trailing underscore, in which case json as character string is returned.

References

API docs at https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/

Examples

head(rcax_nosa())
#>        recoverydomain                                             esu_dps
#> 495 Interior Columbia Salmon, Chinook (Snake River spring/summer-run ESU)
#> 441 Interior Columbia Salmon, Chinook (Snake River spring/summer-run ESU)
#> 3   Interior Columbia Salmon, Chinook (Snake River spring/summer-run ESU)
#> 634 Interior Columbia Salmon, Chinook (Snake River spring/summer-run ESU)
#> 21  Interior Columbia Salmon, Chinook (Snake River spring/summer-run ESU)
#> 387 Interior Columbia Salmon, Chinook (Snake River spring/summer-run ESU)
#>                      commonpopname     commonname    run popid
#> 495 Catherine Creek spring Chinook Chinook salmon Spring     7
#> 441 Catherine Creek spring Chinook Chinook salmon Spring     7
#> 3   Catherine Creek spring Chinook Chinook salmon Spring     7
#> 634 Catherine Creek spring Chinook Chinook salmon Spring     7
#> 21  Catherine Creek spring Chinook Chinook salmon Spring     7
#> 387                          GRCAT Chinook salmon spring     7
#>           majorpopgroup spawningyear nosaij nosaej
#> 495 Grande Ronde/Imnaha         1993    107     99
#> 441 Grande Ronde/Imnaha         1997     72     72
#> 3   Grande Ronde/Imnaha         2005     38     35
#> 634 Grande Ronde/Imnaha         2010    463    438
#> 21  Grande Ronde/Imnaha         2011    615    540
#> 387 Grande Ronde/Imnaha         2013    511