Query the NOSA table

Returns NOSA table sorted by popid and spawningyear. The default columns to return are "recoverydomain", "esu_dps", "commonpopname", "commonname", "run", "popid", "majorpopgroup", "spawningyear", "tsaij", "tsaej", "nosaij", "nosaej"

rcax_nosa() will download 1000 records from the NOSA table. You will probably want to make a filtered query by passing in value to filter on via flist. See rcax_filter(). Examples for the NOSA table:

rcax_nosa(flist=list(popid=7) return values for popid 7.

rcax_nosa(flist=list(esu_dps="Salmon, Chinook (Snake River spring/summer-run ESU)") return values for one ESU. Use rCAX:::caxesu to see a list of ESU_DPS names.

rcax_nosa(
  tablename = "NOSA",
  flist = NULL,
  qlist = NULL,
  cols = c("recoverydomain", "esu_dps", "commonpopname", "commonname", "run", "popid",
    "majorpopgroup", "spawningyear", "tsaij", "tsaej", "nosaij", "nosaej"),
  sortcols = c("spawningyear", "popid"),
  type = c("data.frame", "colnames"),
  GETargs = list(table_id = NULL, recordloc = "records", key = NULL, parse = TRUE),
  ...
)

Arguments

tablename

(character) The CAX name for the table; ignored if GETargs$table_id passed in.

flist

A filter for the query. See details and rcax_filter()

qlist

Additional query params. See rcax_table_query() for details.

cols

column names to return. Use cols=NULL if you want all columns

sortcols

The columns to sort on for the returned table. The order of sortcols indicates the order of sorting. Start with the smallest group and work outwards, e.g. c("spawningyear", "popid", "MPG")

type

whether to return the table ("data.frame") or colnames ("colnames"). Default is to return the table as a data frame.

GETargs

A list of arguments for the rcax_GET() call. These do not normally need to be specified by the user. table_id This is the CAX table id. It is looked up using tablename and caxpops. recordloc This is the name of the record we want in the list returned from the GET call. key is the API key. See rcax_GET() for API key details.

Value

data frame if type="data.frame" (default) and the colnames if type="colnames"

Details

The NOSA table_id is "4EF09E86-2AA8-4C98-A983-A272C2C2C7E3" and is set automatically using a saved data frame from a rcax_tables() call. The table is saved in R/sysdata.rda.

References

API docs at https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/

See also

Examples

a <- rcax_nosa(flist=list(popid=7))
a[, c("popid", "spawningyear", "nosaij", "nosaej")] |> head()
#>    popid spawningyear nosaij nosaej
#> 31     7         1953   4273   3484
#> 46     7         1954    973    816
#> 16     7         1955    122    106
#> 17     7         1956   2606   2338
#> 20     7         1957   1754   1619
#> 13     7         1958    486    456

rcax_nosa(qlist=list(limit=1), cols=NULL)
#> [1] "The table includes no data. Did you specify cols that do not appear in the table? Use rcax_nosa_xport(type=\"colnames\")\" to see the colnames. Names are case insensitive."

# to print the first 5 column names
rcax_nosa(type="colnames")[1:5]
#> [1] "nosaijupperlimit"   "tsaej"              "age8propupperlimit"
#> [4] "metriclocation"     "popfitnotes"