Return NOSA base table

Returns the base HLI table which is missing the additional metadata (such as NMFS_PopID and update dates) found in the XPort tables.

The query will download 1000 records by default; you can change this by passing in qlist=list(limit=1100) for example. To make a filtered query e.g. for one ESU or popid) pass in the flist argument. Field names and values are case insensitive, e.g. POPID, popid and PopID return the same result. If your filter value is not in the table, the returned table will be empty. Examples:

flist=list(popid=7) return values for popid 7.

flist=list(esu_dps="Salmon, Chinook (Snake River spring/summer-run ESU)") return values for one ESU. Use rCAX:::caxesu to see a list of ESU_DPS names.

rcax_nosa(
  tablename = "NOSA",
  flist = NULL,
  qlist = NULL,
  cols = NULL,
  sortcols = c("spawningyear", "popid"),
  type = c("data.frame", "colnames"),
  GETargs = list(table_id = NULL, recordloc = "records", key = NULL, parse = TRUE),
  ...
)

Arguments

tablename

(character) The CAX name for the table; ignored if GETargs$table_id passed in.

flist

A filter for the query. See details and rcax_filter()

qlist

Additional query params. See rcax_table_query() for details.

cols

column names to return. Use cols=NULL if you want all columns

sortcols

The columns to sort on for the returned table. The order of sortcols indicates the order of sorting. Start with the smallest group and work outwards, e.g. c("spawningyear", "popid", "MPG")

type

whether to return the table ("data.frame") or colnames ("colnames"). Default is to return the table as a data frame.

GETargs

A list of arguments for the rcax_GET() call. These do not normally need to be specified by the user. table_id This is the CAX table id. It is looked up using tablename and caxpops. recordloc This is the name of the record we want in the list returned from the GET call. key is the API key. See rcax_GET() for API key details.

Value

data frame if type="data.frame" (default) and the colnames if type="colnames"

Details

The table_id is set automatically using a saved data frame from a rcax_tables() call. The table is saved in R/sysdata.rda.

If you want to see only the column names use type="colnames.

References

API docs at https://www.streamnet.org/resources/exchange-tools/rest-api-documentation/

See also

Examples

# return popid of first record
# Note the part after $ is case sensitive
id <- rcax_nosa(qlist=list(limit=1))$popid

a <- rcax_nosa(flist=list(popid=id))
a <- a[, c("popid", "spawningyear", "nosaij", "nosaej")]
head(a)
#>   popid spawningyear nosaij nosaej
#> 1    58         1990    922       
#> 2    58         1991   1230       
#> 3    58         1992    739       
#> 4    58         1993    335       
#> 5    58         1994    318       
#> 6    58         1995    460       

# to print the first 5 column names
rcax_nosa(type="colnames")[1:5]
#> [1] "nosaijupperlimit"   "tsaej"              "age8propupperlimit"
#> [4] "metriclocation"     "popfitnotes"